Started work on a new configuration display.
In my on going efforts to make the process of configuring and
editing the content presented by the CMSContent portlet I have
started another round of UI redesign. This time i'm trying to make
it feel as inline as possible, as if they were editing the thing
live. They basically will be but with some magic to fill in the
gaps. This commit sees also the initial move to web components
using vue. So far I have added the config-bar and the layout
selector for said bar. It has a little bit of tweaking left to do
to get it fully operational, but that component is almost done.