Further work on new config screen.
I've fine tuned the layout selector and added the document source
selector and a place holder document selector. The document
selector is going to be the most complicated as I'm leaning towards
making it a tree view in a drop down with search. I want to further
re-enforce the analogy of document ids being document paths. Having
the document selector as a tree view aids that mission.

There is still a lot to do to get it to the shape I imagine, but it
is coming along.