I have upgraded reports to use web components.
It took me a bit of work and an additional API but I have changed
the way reports are displayed to have it load the content
dynamically instead of having it all pre populated. This was a
necessity for reports that have a large number of responses. Once
a report reaches about 300 responses or so the render time would
get to high and the page wouldn't respond.