Fixed issues with loading.
There were two issues that persisted from yesterday. One was that
the document was loading only the first 25 and this was due to a
issue where the reponse to a http request was not effecting the
vue object proper. So the row numbers weren't getting set correctly
in said object. The other issue was double loading. This was caused
by a miss understanding of hibernate. I adjusted my logic to
properly set the maximum results.

Now I just need to finish the token work.