Tokenizing complete.
I have added security tokens to the report api to keep senative
information safe from prying eyes. The system is not perfect as a
person could try to sneak a token away from someone who has
legimate access to a report. But since it's https that attack
surface is very small. We are getting into the realm of social
engineering to achieve that. And if they can do that, they can just
use the credentials they stole to see the reports, screw my tokens.
I have provided a mechanism for limiting what tokens can access by
providing a means for the code making the token to state what it is
for. Then when the token is checked for validity, if the request is
for what the code said the request would be for matches then it is
validated.