Got XLSX output working.
There is a little futher work I need to do to secure it and to get
a link to the excel download into the report. I will probably also
want to clean up the code a bit but for now I have a attachment
portlet fire that needs to be put out.