Translation steps in progress.
I've got the code in place so that documents can be of either the
new or old ID type and still work correclty. At least in the config
and display screens. The editor will need a fair bit of love to be
back in a good way.

I also added a CMSId_map class which maps between the new and old
ids. It may have to be updated as I work on this but that won't be
to hard.