Rest v1 has been removed!
I have done it! I've gotten all of these references to the old
rest API fully removed. So now the portlet runs exclusively on the
new API. The new API needs some refinement of it's own yet,
specifically the report and form APIs need to have the whole nouns
only treatement, and some of the document API still needs the same.
However it is always nice to cull old code.

Some work on the editor that still needs to be done: I've got to
be sure that save is working correctly, and now I also need to
provide a mechanism for moving documents in the tree.