Got it! Move is now functioning.
You can now move documents around without changing subscriptions.
This fulfils the feature request that marketing had had for me. I
still need to implement this in the form editor, but it is great
to have it done in at least one. That means I've done the hard part
now I just need to, unfortunately, copy paste everything. I will
hopefully be doing that UI rewrite this spring and then I won't
have seperate systems for documents and forms. That would be nice.