Report download now fully functioning.
I need to merge this into the main line now and continue working on
it. In the branch 'document_path' I have done a lot of revision to
the API system and this new feature makes use of it. So I'm looking
at a pretty messy merge I'm fairly certain. But that is the next
step for this. I've got a download button that forms the URL for
the document correctly. The one thing I can't seem to get to work
is the excel icon in the download button.