Form editor with components in.
I've started to rework some of the CMSForm editor. Specifically I
am replacing the form fields part of that system. I'm turning it
into a pair of components, atleast so far only a pair. I've made
great strides today and have something that already saves. At least
the basic idea does. required and the options field are a little
broken, but for the most part it works. I'm getting the options
field to be up to snuff right now. Then I'll polish the rest and
make sure that it works with all the other existing systems.