Time to move to dev.
I'm getting some odd issues related to the database. It should be
connecting correctly to the dev database from my machine, but it
doesn't seem to be. So I'm going to move this process over to dev
now, as I was going to have to eventually. I'll continue work there
getting the load and save functions fully back to snuff.