Frustration frustration!
I've gotten loading working correctly after much headache around
getting the frakin list to clear of it's old items. But now I'm
stuck on a similar problem. For some reason I cannot for the life
of me get the remove function to work. I have followed the guide
on how this works and it fails. I have even gone in and figured out
how to do it manually and watched as the wrong items were removed.
It is always the last item regardless of what I do. I couldn't tell
you why. If I could I might be able to solve this problem. >.<