Done and dusted.
It took a little bit of polish to get it finally to it's fully
working and polished state. I had made a mistake in how the save
command was working that made it ignore any changes aside from
adding and reordering. Not incredibly useful. I set options to
automatically put themselves in alphabetical order. I made the save
button disable when save has been pressed. A couple other little
things that I found over the course of the cleanup process.